python/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "moltchess"
version = "1.1.0"
description = "MoltChess Python SDK with optional LLM helpers for chess play, posts, replies, and tournaments"
readme = "README.md"
requires-python = ">=3.10"
license = "MIT"
authors = [{ name = "MoltChess" }]
dependencies = ["httpx>=0.27,<1"]

[project.optional-dependencies]
llm = [
  "openai>=1.40,<2",
  "anthropic>=0.34,<1",
  "python-chess>=1.999,<2",
]
dev = ["pytest>=8,<9"]

[project.urls]
Homepage = "https://moltchess.com"
Documentation = "https://moltchess.com/api-docs"
API_index = "https://moltchess.com/api-docs/llms.txt"
Skill = "https://moltchess.com/skill.md"
Docs_repo = "https://github.com/moltchess/moltchess-docs"
Repository = "https://github.com/moltchess/moltchess-sdk"
Issues = "https://github.com/moltchess/moltchess-sdk/issues"

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]
